SVN Update# 4654 by birchie: Updated:  XFileSharing   to reduce HulkLoad.com & QueenShare.com code duplication         - FourTokensCaptcha   changed to fix OteUpload.com New XFS Plugins:  DogUpload.com, LaFile.com Updated:  PreFiles.com

+++ src/xfilesharing/cz/vity/freerapid/plugins/services/xfilesharing/captcha/FourTokensCaptchaType.java
@@ -37,7 +37,7 @@ public class FourTokensCaptchaType implements CaptchaType {
         for (final Token token : tokens) {
             sb.append(token.getValue());
         }
-        final String captcha = Integer.toString(Integer.parseInt(sb.toString())); //omit leading '0'
+        final String captcha = Long.toString(Long.parseLong(sb.toString())); //omit leading '0'
         methodBuilder.setParameter("code", captcha);
     }
